+++ src/phonometry/emission/intensity.py
@@ -47,9 +47,9 @@ if TYPE_CHECKING:
     from matplotlib.axes import Axes
 from scipy import signal
 
-from ._internal.levels_math import energy_mean
-from .metrology.frequencies import _genfreqs
-from ._internal.utils import _typesignal
+from .._internal.levels_math import energy_mean
+from ..metrology.frequencies import _genfreqs
+from .._internal.utils import _typesignal
 
 #: Reference sound intensity, in watts per square metre (ISO 9614-1, A.2.3).
 _I0 = 1.0e-12
@@ -118,7 +118,7 @@ class IntensityResult:
         and matplotlib (``pip install phonometry[plot]``); returns the
         :class:`~matplotlib.axes.Axes`.
         """
-        from ._plotting import plot_intensity
+        from .._plot.emission import plot_intensity
 
         return plot_intensity(self, ax=ax, **kwargs)
