+++ src/phonometry/emission/sound_power_reverberation.py
@@ -47,7 +47,7 @@ import numpy as np
 if TYPE_CHECKING:
     from matplotlib.axes import Axes
 
-from ._internal.levels_math import energy_mean, energy_sum
+from .._internal.levels_math import energy_mean, energy_sum
 from .sound_power import (
     SoundPowerWarning,
     _a_weighting_corrections,
@@ -96,7 +96,7 @@ class ReverberationSoundPowerResult:
         Requires matplotlib (``pip install phonometry[plot]``); returns the
         :class:`~matplotlib.axes.Axes`.
         """
-        from ._plotting import plot_sound_power
+        from .._plot.emission import plot_sound_power
 
         return plot_sound_power(self, ax=ax, **kwargs)
