+++ src/phonometry/hearing/__init__.py
@@ -1,0 +1,47 @@
+#  Copyright (c) 2026. Jose M. Requena-Plens
+"""hearing domain of phonometry (see module docstrings)."""
+
+from __future__ import annotations
+
+from .noise_induced_hearing_loss import HtlanResult, NiptsResult, htlan, nipts
+from .occupational_exposure import (
+    COVERAGE_FACTOR,
+    ExposureResult,
+    INSTRUMENT_U2,
+    OccupationalExposureWarning,
+    Task,
+    TaskContribution,
+    full_day_exposure,
+    job_based_exposure,
+    minimum_cumulative_duration_hours,
+    table_c4_contribution,
+    task_based_exposure,
+)
+from .sii import SIIResult, speech_intelligibility_index, standard_speech_spectrum
+from .sti import STIResult, STIWarning, sti_from_impulse_response, stipa, stipa_signal
+
+__all__ = [
+    "COVERAGE_FACTOR",
+    "ExposureResult",
+    "HtlanResult",
+    "INSTRUMENT_U2",
+    "NiptsResult",
+    "OccupationalExposureWarning",
+    "SIIResult",
+    "STIResult",
+    "STIWarning",
+    "Task",
+    "TaskContribution",
+    "full_day_exposure",
+    "htlan",
+    "job_based_exposure",
+    "minimum_cumulative_duration_hours",
+    "nipts",
+    "speech_intelligibility_index",
+    "standard_speech_spectrum",
+    "sti_from_impulse_response",
+    "stipa",
+    "stipa_signal",
+    "table_c4_contribution",
+    "task_based_exposure",
+]
