+++ tests/emission/test_intensity.py
