+++ tests/emission/test_sound_power.py
