+++ tests/emission/test_sound_power_reverberation.py
