+++ tests/emission/test_vibration_sound_power.py
